database/requirements.txt
pocketbase==0.14.0
python-dotenv==1.0.1
Requests==2.32.3
